Remove support for "realtime" updates.
This was half-baked, and ran the script every frame, forcibly, even if the next run was seconds or minutes away. The game does a bunch of expensive prep work to create permissions-handling proxies for objects on connected grids every time a script runs, so this imposed a bunch of additional per-frame load for each TIM in the world. No bueno, and I'm not actually using it, so it has to go.